+++ otelconf/x/config_json.go
@@ -123,6 +123,21 @@ func (j *TraceContextPropagator) UnmarshalJSON(b []byte) error {
 }
 
 // UnmarshalJSON implements json.Unmarshaler.
+func (j *ExperimentalAzureVMResourceDetector) UnmarshalJSON(b []byte) error {
+	type plain ExperimentalAzureVMResourceDetector
+	var p plain
+	if err := json.Unmarshal(b, &p); err != nil {
+		return errors.Join(newErrUnmarshal(j), err)
+	}
+	if p == nil {
+		*j = ExperimentalAzureVMResourceDetector{}
+	} else {
+		*j = ExperimentalAzureVMResourceDetector(p)
+	}
+	return nil
+}
+
+// UnmarshalJSON implements json.Unmarshaler.
 func (j *ExperimentalContainerResourceDetector) UnmarshalJSON(b []byte) error {
 	type plain ExperimentalContainerResourceDetector
 	var p plain
@@ -192,6 +207,7 @@ func (j *ExperimentalResourceDetector) UnmarshalJSON(b []byte) error {
 	type Plain ExperimentalResourceDetector
 	type shadow struct {
 		Plain
+		AzureVM   json.RawMessage `json:"azure.vm"`
 		Container json.RawMessage `json:"container"`
 		Host      json.RawMessage `json:"host"`
 		Process   json.RawMessage `json:"process"`
@@ -200,6 +216,17 @@ func (j *ExperimentalResourceDetector) UnmarshalJSON(b []byte) error {
 	var sh shadow
 	if err := json.Unmarshal(b, &sh); err != nil {
 		return errors.Join(newErrUnmarshal(j), err)
+	}
+
+	if sh.AzureVM != nil {
+		var c ExperimentalAzureVMResourceDetector
+		if err := json.Unmarshal(sh.AzureVM, &c); err != nil {
+			return errors.Join(newErrUnmarshal(j), err)
+		}
+		if c == nil {
+			c = ExperimentalAzureVMResourceDetector{}
+		}
+		sh.Plain.AzureVM = c
 	}
 
 	if sh.Container != nil {